LUT in progress

+++ src/spiky/lut/tests/test_lut_backward.py
@@ -125,7 +125,7 @@ def main():
         devices.append('cuda')
 
     for device in devices:
-        for summation_dtype in [torch.float32]:  # , torch.int32
+        for summation_dtype in [torch.int32]:  # , torch.float32
             print(f"\nTesting on {device}, summation_dtype {summation_dtype}...")
             success = test_lut_backward(device, summation_dtype)
